+++ pkg/lmap/common.go
