+++ pkg/lmap/ping.go
